xen: credit2: soft-affinity awareness in fallback_cpu()
By, basically, moving all the logic of the function
inside the usual two steps (soft-affinity step and
hard-affinity step) loop.

Signed-off-by: Dario Faggioli <dario.faggioli@citrix.com>
Signed-off-by: Justin T. Weaver <jtweaver@hawaii.edu>
---
Cc: Anshul Makkar <anshul.makkar@citrix.com>
---
George, you gave your Reviewed-by to:
 https://lists.xenproject.org/archives/html/xen-devel/2016-08/msg02201.html

which was adding soft-affinity awareness to both fallback_cpu and cpu_pick(). See here:
 https://lists.xenproject.org/archives/html/xen-devel/2016-09/msg03259.html

I changed the cpu_pick() part a lot, and that's why I decided to split the
patch in two.  As far as fallback_cpu(), though, what's done in this patch is
exactly the same that was being done in the original one.

So, of course I'm dropping the Rev-by, but I thought it could have been useful
to mention this. :-)